refactor: separate schema from compiler internals
