+++ packages/core/src/adapters/mysql2.ts
@@ -1,7 +1,10 @@
 import type { Pool, Connection } from 'mysql2/promise';
 import type { ExecuteValues } from 'mysql2';
-import type { DialectExecutor, QueryMeta, SchemaLike, TypedDb, TypedDbOptions } from '../index.js';
-import { createTypedDb } from '../index.js';
+import type { TypedDb, TypedDbOptions } from '../public/client.js';
+import { createTypedDb } from '../public/client.js';
+import type { DialectExecutor } from '../runtime/executor.js';
+import type { QueryMeta } from '../runtime/result.js';
+import type { SchemaLike } from '../schema/model.js';
 import { rollbackAndRethrow } from './transaction.js';
 
 export type Mysql2Queryable = Pool | Connection;
