+++ packages/core/src/adapters/pg.ts
@@ -1,6 +1,8 @@
 import type { Pool, Client, PoolClient } from 'pg';
-import type { DialectExecutor, SchemaLike, TypedDb, TypedDbOptions } from '../index.js';
-import { createTypedDb } from '../index.js';
+import type { TypedDb, TypedDbOptions } from '../public/client.js';
+import { createTypedDb } from '../public/client.js';
+import type { DialectExecutor } from '../runtime/executor.js';
+import type { SchemaLike } from '../schema/model.js';
 import { rollbackAndRethrow } from './transaction.js';
 
 export type PgQueryable = Pool | Client | PoolClient;
