+++ packages/core/src/public/query.ts
@@ -5,7 +5,7 @@ import type {
   InferStrictViaGateway,
   InferViaGateway,
 } from '../compiler/gateway.js';
-import type { SchemaLike } from '../compiler/schema/model.js';
+import type { SchemaLike } from '../schema/model.js';
 
 export type {
   InferParamsViaGateway as InferParams,
